requirements.txt
torch # pip install torch --extra-index-url https://download.pytorch.org/whl/cu113
pytorch-lightning
pytorch_msssim
monai
torchmetrics
torch-fidelity
torchio 
pillow
einops
torchvision
matplotlib
pandas 
lpips

streamlit